+++ routes/api.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\AuthController;
 use App\Http\Controllers\Api\SchoolController;
 use App\Http\Controllers\Api\StudentController;
 use App\Http\Controllers\Api\AttendanceController;
+use App\Http\Controllers\Api\AttendanceExportController;
 
 Route::post('/register', [AuthController::class, 'register']);
 Route::post('/login', [AuthController::class, 'login']);
@@ -26,3 +27,6 @@ Route::get('/attendances', [AttendanceController::class, 'index']);
 Route::post('/attendances/{fingerprint_id}', [AttendanceController::class, 'store']);
 Route::get('/attendances/{id}', [AttendanceController::class, 'show']);
 Route::delete('/attendances/{id}', [AttendanceController::class, 'destroy']);
+
+// Attendance export route
+Route::get('/export', [AttendanceExportController::class, 'export']);