+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 - Add Schools Endpoint
 - Add Students Endpoint
 - Add Attendances Endpoint
+- Add Export Endpoint